add helper script to generate rails specfile templates
this helps with the typical rails packaging and is currently used
by portus, and will be used by kubic-project/velum as well to help
with the suse specific packaging.

for not having to maintain the same code in 2 different repos i'd
like to add it to the general suse ruby-packaging repo and reference
it in portus and velum from there.